Create liquidTestRunner file
