+++ lib/cli/devMode.js
@@ -1,6 +1,7 @@
 const fs = require("fs");
 const path = require("path");
 const toolkit = require("../../index");
+const liquidTestRunner = require("../liquidTestRunner");
 const fsUtils = require("../utils/fsUtils");
 const CWD = process.cwd();
 
@@ -36,7 +37,12 @@ async function watchLiquidTest(firmId, handle, testName, html_render) {
     lastUpdates[filename] = fileStats.mtimeMs;
 
     // Run test
-    await toolkit.runTestsWithOutput(firmId, handle, testName, html_render);
+    await liquidTestRunner.runTestsWithOutput(
+      firmId,
+      handle,
+      testName,
+      html_render
+    );
   });
 
   // Watch liquid files
@@ -51,7 +57,12 @@ async function watchLiquidTest(firmId, handle, testName, html_render) {
       lastUpdates[filename] = fileStats.mtimeMs;
 
       // Run test
-      await toolkit.runTestsWithOutput(firmId, handle, testName, html_render);
+      await liquidTestRunner.runTestsWithOutput(
+        firmId,
+        handle,
+        testName,
+        html_render
+      );
     });
   });
 }
